+++ montepy/input_parser/__init__.py
@@ -10,4 +10,6 @@ from . import parser_base
 from . import read_parser
 from . import shortcuts
 from . import surface_parser
+from . import tally_parser
+from . import tally_seg_parser
 from . import tokens