+++ doc/source/changelog.rst
@@ -2,11 +2,16 @@ MontePy Changelog
 =================
 
 #Next Version#
----------------------
+-------------------
 
 **Features Added**
 
 * ``overwrite`` argument added to `MCNP_Problem.write_to_file` to ensure files are only overwritten if the user really wants to do so (:pull:`443`).
+
+**Bug fixes**
+
+* Fixed bug with ``SDEF`` input, and made parser more robust (:issue:`396`).
+
 
 0.2.10
 ----------------------